Refactoring of Bubble classes.
Changed some names, added some JavaDoc and fixed a minor bug with the root nodes.
