+++ PL2/PL2-core/src/main/java/nl/tudelft/pl2016gr2/core/algorithms/subgraph/CompareSubgraphs.java
@@ -168,7 +168,7 @@ public class CompareSubgraphs {
     areaMap.put(node, complexNodeArea);
 
     SimpleVerticalArea nodeArea = complexNodeArea.getLargestArea();
-    
+
     node.getGuiData().relativeYPos = nodeArea.getCenter() / VERTICAL_PRECISION;
     node.getGuiData().maxHeight = nodeArea.getHeight() / (double) VERTICAL_PRECISION;
     return complexNodeArea;
