+++ PL2/PL2-core/src/main/java/nl/tudelft/pl2016gr2/core/algorithms/subgraph/SubgraphAlgorithmManager.java
@@ -1,6 +1,6 @@
 package nl.tudelft.pl2016gr2.core.algorithms.subgraph;
 
-import nl.tudelft.pl2016gr2.core.algorithms.FilterPhyloBubbles;
+import nl.tudelft.pl2016gr2.core.algorithms.PhyloBubbleFilter;
 import nl.tudelft.pl2016gr2.core.algorithms.mutations.MutationBubbleAlgorithms;
 import nl.tudelft.pl2016gr2.model.graph.SequenceGraph;
 import nl.tudelft.pl2016gr2.model.graph.data.BaseSequence;
@@ -105,7 +105,7 @@ public class SubgraphAlgorithmManager {
       return first.getLevel() - second.getLevel();
     });
 
-    FilterPhyloBubbles filter = new FilterPhyloBubbles(orderedNodes);
+    PhyloBubbleFilter filter = new PhyloBubbleFilter(orderedNodes);
     orderedNodes = filter.filter(treeRoot, genomes);
     CompareSubgraphs.alignVertically(orderedNodes);
 
@@ -183,7 +183,7 @@ public class SubgraphAlgorithmManager {
     public void run() {
       orderedNodes = subgraph.getOrderedGraph();
       orderedNodes = MutationBubbleAlgorithms.makeBubbels(orderedNodes);
-      FilterPhyloBubbles filter = new FilterPhyloBubbles(orderedNodes);
+      PhyloBubbleFilter filter = new PhyloBubbleFilter(orderedNodes);
       orderedNodes = filter.filter(treeRoot, genomes);
       ArrayList<GraphNode> newNodes = new ArrayList<>();
       for (GraphNode node : orderedNodes) {
