+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/model/graph/nodes/AbstractGraphNode.java
@@ -154,7 +154,7 @@ public abstract class AbstractGraphNode implements GraphNode {
     //inEdges.trimToSize();
     //outEdges.trimToSize();
   }
-  
+
   @Override
   public String toString() {
     return "id: " + getId();
