+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/model/graph/nodes/GraphBubble.java
@@ -1,112 +1,78 @@
 package nl.tudelft.pl2016gr2.model.graph.nodes;
 
+import nl.tudelft.pl2016gr2.visitor.NodeVisitor;
+
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 
-public abstract class GraphBubble extends Bubble {
-  
-  private final PhyloFilter filter;
-  private List<GraphNode> poppedNodes;
-  private boolean isPopped;
+public class GraphBubble extends AbstractGraphBubble {
 
-  private final HashMap<Integer, Collection<GraphNode>> originalOutEdges = new HashMap<>(4);
-  private final HashMap<Integer, Collection<GraphNode>> originalInEdges = new HashMap<>(4);
-  private final HashMap<Integer, Collection<GraphNode>> unpoppedOutEdges = new HashMap<>(4);
-  private final HashMap<Integer, Collection<GraphNode>> unpoppedInEdges = new HashMap<>(4);
-  
-  public GraphBubble(int id, PhyloFilter filter) {
-    super(id);
-    this.filter = filter;
-  }
-  
-  public GraphBubble(int id, PhyloFilter filter,
-      Collection<GraphNode> inEdges, Collection<GraphNode> outEdges) {
-    super(id, inEdges, outEdges);
-    this.filter = filter;
-  }
-  
-  public GraphBubble(int id, PhyloFilter filter,
-      Collection<GraphNode> inEdges, Collection<GraphNode> outEdges, List<GraphNode> nestedNodes) {
-    super(id, inEdges, outEdges, nestedNodes);
-    this.filter = filter;
-  }
-  
-  protected GraphBubble(Bubble bubble, PhyloFilter filter) {
-    super(bubble);
-    this.filter = filter;
-  }
-  
-  @Override
-  public int getGenomeSize() {
-    int count = 0;
-    for (GraphNode inEdge : getInEdges()) {
-      count += inEdge.getGenomeSize();
-    }
-    return count;
-  }
-  
   /**
-   * Returns the phylofilter of this bubble.
-   * 
-   * @return an instance of PhyloFilter.
+   * Construct a graph bubble.
+   *
+   * @param id       the id of the bubble.
+   * @param filter   the filter object to call when zooming in.
    */
-  protected PhyloFilter getFilter() {
-    return filter;
+  public GraphBubble(int id, BubbleFilter filter) {
+    super(id, filter);
+  }
+
+  /**
+   * Construct a graph bubble.
+   *
+   * @param id       the id of the bubble.
+   * @param filter   the filter object to call when zooming in.
+   * @param inEdges  the in edges of the bubble.
+   * @param outEdges the out edges of the bubble.
+   */
+  public GraphBubble(int id, BubbleFilter filter,
+      Collection<GraphNode> inEdges, Collection<GraphNode> outEdges) {
+    super(id, filter, inEdges, outEdges);
+  }
+
+  /**
+   * Construct a graph bubble.
+   *
+   * @param id          the id of the bubble.
+   * @param filter      the filter object to call when zooming in.
+   * @param inEdges     the in edges of the bubble.
+   * @param outEdges    the out edges of the bubble.
+   * @param nestedNodes the nested nodes of the bubble.
+   */
+  public GraphBubble(int id, BubbleFilter filter,
+      Collection<GraphNode> inEdges, Collection<GraphNode> outEdges, List<GraphNode> nestedNodes) {
+    super(id, filter, inEdges, outEdges, nestedNodes);
+  }
+
+  /**
+   * Copy a graph bubble.
+   *
+   * @param bubble   the super class of the phylo bubble to copy.
+   * @param filter   the filter object to call when zooming in.
+   */
+  private GraphBubble(Bubble bubble, BubbleFilter filter) {
+    super(bubble, filter);
+  }
+
+  @Override
+  public GraphNode copy() {
+    return new GraphBubble(getId(), getFilter());
+  }
+
+  @Override
+  public GraphNode copyAll() {
+    return new GraphBubble(this, getFilter());
+  }
+
+  @Override
+  public String toString() {
+    return String.format("%s: \n%s", "Random Bubble", super.toString());
   }
   
   @Override
-  public boolean needsVerticalAligning() {
-    return poppedNodes == null;
+  public void accept(NodeVisitor visitor) {
+    visitor.visit(this);
   }
-  
-  @Override
-  public Collection<GraphNode> pop() {
-    if (!isPopped) {
-      isPopped = true;
-      if (poppedNodes == null) {
-        for (GraphNode node : getInEdges()) {
-          originalOutEdges.put(node.getId(), new HashSet<>(node.getOutEdges()));
-        }
-        for (GraphNode node : getOutEdges()) {
-          originalInEdges.put(node.getId(), new HashSet<>(node.getInEdges()));
-        }
-        poppedNodes = filter.zoomIn(this);
-      } else {
-        for (GraphNode node : getInEdges()) {
-          node.setOutEdges(unpoppedOutEdges.get(node.getId()));
-        }
-        for (GraphNode node : getOutEdges()) {
-          node.setInEdges(unpoppedInEdges.get(node.getId()));
-        }
-      }
-    }
-    return poppedNodes;
-  }
-
-  @Override
-  public void unpop() {
-    if (isPopped) {
-      isPopped = false;
-      for (GraphNode node : getInEdges()) {
-        unpoppedOutEdges.put(node.getId(), new HashSet<>(node.getOutEdges()));
-      }
-      for (GraphNode node : getOutEdges()) {
-        unpoppedInEdges.put(node.getId(), new HashSet<>(node.getInEdges()));
-      }
-      for (GraphNode node : getInEdges()) {
-        node.setOutEdges(originalOutEdges.get(node.getId()));
-      }
-      for (GraphNode node : getOutEdges()) {
-        node.setInEdges(originalInEdges.get(node.getId()));
-      }
-    }
-  }
-
-  @Override
-  public boolean isPopped() {
-    return isPopped;
-  }
-
 }
+
+
