+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/model/graph/nodes/PhyloBubble.java
@@ -6,7 +6,7 @@ import nl.tudelft.pl2016gr2.visitor.NodeVisitor;
 import java.util.Collection;
 import java.util.List;
 
-public class PhyloBubble extends GraphBubble {
+public class PhyloBubble extends AbstractGraphBubble {
 
   private final IPhylogeneticTreeNode treeNode;
 
@@ -17,7 +17,7 @@ public class PhyloBubble extends GraphBubble {
    * @param treeNode the phylogenetic tree node where the mutation of this bubble occurs.
    * @param filter   the filter object to call when zooming in.
    */
-  public PhyloBubble(int id, IPhylogeneticTreeNode treeNode, PhyloFilter filter) {
+  public PhyloBubble(int id, IPhylogeneticTreeNode treeNode, BubbleFilter filter) {
     super(id, filter);
     this.treeNode = treeNode;
   }
@@ -31,7 +31,7 @@ public class PhyloBubble extends GraphBubble {
    * @param inEdges  the in edges of the bubble.
    * @param outEdges the out edges of the bubble.
    */
-  public PhyloBubble(int id, IPhylogeneticTreeNode treeNode, PhyloFilter filter,
+  public PhyloBubble(int id, IPhylogeneticTreeNode treeNode, BubbleFilter filter,
       Collection<GraphNode> inEdges, Collection<GraphNode> outEdges) {
     super(id, filter, inEdges, outEdges);
     this.treeNode = treeNode;
@@ -47,7 +47,7 @@ public class PhyloBubble extends GraphBubble {
    * @param outEdges    the out edges of the bubble.
    * @param nestedNodes the nested nodes of the bubble.
    */
-  public PhyloBubble(int id, IPhylogeneticTreeNode treeNode, PhyloFilter filter,
+  public PhyloBubble(int id, IPhylogeneticTreeNode treeNode, BubbleFilter filter,
       Collection<GraphNode> inEdges, Collection<GraphNode> outEdges, List<GraphNode> nestedNodes) {
     super(id, filter, inEdges, outEdges, nestedNodes);
     this.treeNode = treeNode;
@@ -60,7 +60,7 @@ public class PhyloBubble extends GraphBubble {
    * @param treeNode the phylogenetic tree node where the mutation of this bubble occurs.
    * @param filter   the filter object to call when zooming in.
    */
-  private PhyloBubble(Bubble bubble, IPhylogeneticTreeNode treeNode, PhyloFilter filter) {
+  private PhyloBubble(Bubble bubble, IPhylogeneticTreeNode treeNode, BubbleFilter filter) {
     super(bubble, filter);
     this.treeNode = treeNode;
   }
